cgroup: enable only the controllers needed
only the root user can enable all the controllers for the unprivileged
users.  Change it to request only the controllers we need.
In the cgroup v2 unified mode, systemd enables by default the pids and
memory controllers, an unprivileged user can use them without any
change to the system configuration.

Signed-off-by: Giuseppe Scrivano <giuseppe@scrivano.org>