utils: fix file existence check
simplify logic for crun_path_exists to not check for the file mode.
It was previously checking whether the file is either writeable or
readable.

This caused an issue with /proc/sysrq-trigger when running in a user
namespace, as the file that has mode 0200 is not owned by root in the
user namespace so crun_path_exists(F_READ) would report it as not
existing and skip to create a bind mount and make it read-only.

Closes: https://github.com/containers/crun/issues/74

Signed-off-by: Giuseppe Scrivano <giuseppe@scrivano.org>