seccomp: move generation out of cgroup
generate the seccomp profile in the parent process, not from init
while running with cgroups configured.

The reason is that seccomp generation might require more memory than
it is configured in the cgroups.

crun memory consumption is more predictable now, and can run with less
than 500Kb of memory in the cgroup (I've tested that it runs reliably
with 250Kb and never hit any OOM).

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>