container: do not fork on create
do not fork the current process on create, this prevents creating
additional processes to wait for.

It solves a race when the OCI runtime caller has SUBREAPER set, such
as Buildah.  When the initial process exited, the container process
cannot be directly waited for as there are other processes to reap
first.

Closes: https://github.com/containers/crun/issues/215

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>