linux: use temporary mount
use a temporary mount location that is not under the rootfs for
setting up mounts.

The temporary mount is used to solve a race condition where the mount
point we created on top of volumes that are accessible also from other
containers.  The temporary mount once configured is moved to its
destination under the rootfs.

Signed-off-by: Giuseppe Scrivano <giuseppe@scrivano.org>