linux: enforce /proc and /sys to be dir under rootfs
enforce /proc and /sys to be directories without any symlinks and to
be direct child of the rootfs.

It solves a race condition when /proc or /sys might be mounted on a
volume where another container has also write access, and successive
mounts of masked and read only paths can be performed on a different
path.

To lift this restriction, it might be necessary to rewrite
maskedPaths/readOnlyPaths relative to /proc and /sys and use the dirfd
pointing to /proc or /sys to do the mount.

Signed-off-by: Giuseppe Scrivano <giuseppe@scrivano.org>