linux: fix the order of clone syscall arguments on s390 and cris
When using podman on s390 (and, presumably, cris), the following error
occurs:

    writing file '/proc/{}/gid_map': Operation not permitted

The root cause is that the order of clone syscall arguments is wrong: on
s390 and cris they are reversed (see "C library/kernel differences"
section of `man 2 clone`), but crun uses the same order for all
architectures. The end result is that the container process is created
with stack == CLONE_NEWUSER and flags == 0, causing it to inherit its
parent's user namespace (which prevents writes to gid_map) and to almost
immediately segfault.

Fix by special-casing clone call sequence for s390 and cris.

See also: https://github.com/containers/libpod/commit/ca5114faf928