linux: re-open /dev/null inside of the container
This is in preparation for the crun checkpoint/restore support.

Checkpointing a container with CRIU fails if the containers stdin,
stdout and stderr has been redirected to /dev/null, but /dev/null still
points to the device outside of the container. With this change stdin,
stdout and stderr will be pointing to /dev/null on the inside of the
container.

With this change it is possible to checkpoint a crun container.

Signed-off-by: Adrian Reber <areber@redhat.com>