cgroup: support systemd properties via annotations
add support for systemd properties passed through annotations in the
form:

"annotations": {
               "org.systemd.property.TimeoutStopUSec": "uint64 123456789",
               "org.systemd.property.CollectMode":"'inactive-or-failed'"
}

add a basic parser for the gvariant types to avoid a dependency on
glib.  The parser doesn't support complex types but it is just enough
for the types accepted by systemd.

Similar to https://github.com/opencontainers/runc/pull/2224

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>