Add initial checkpoint support
This adds the initial support to checkpoint a container using CRIU.
This does not yet add support to restore a container. The goal is to be
able to submit the checkpoint/restore feature in smaller pull requests.

To not break container engines which rely on checkpoint/restore support
detection by running the 'checkpoint' sub-command. The checkpoint
command is still called '_checkpoint' and not shown in the global help.

The checkpoint support is still experimental especially as it not really
useful without restore support.

This tries to implement the features necessary to enable Podman to
checkpoint and restore crun based containers.

Signed-off-by: Adrian Reber <areber@redhat.com>