criu: implement container restore
This is the first very limited implementation to checkpoint and restore
a container using crun via CRIU.

It does not work with a user namespace (yet) and there is no console
handling implemented (in addition to many other missing pieces). But it
enables to have the basic infrastructure in crun to checkpoint and
restore a simple container.

Signed-off-by: Adrian Reber <areber@redhat.com>