tests: add checkpoint restore test
This adds a test to checkpoint and restore a container. This has the
limitations that user namespaces are not yet handled correctly and all
console input and output is not yet set up. There can also be no cgroup2
mounted in the container because CRIU cannot yet handle cgroup2 in the
container.

The new checkpoint restore test only runs if crun has been built with
CRIU support.

Signed-off-by: Adrian Reber <areber@redhat.com>