status: Use process start time for identification
In this patch is introduced a process-start-time status value for
containers. This value, in addition to the PID value, is used to
identify a container process.

This is analogous to the init_process_start state parameter used
by runc.

For backwards compatibility, ignore the check for process start time
when this value is not set in the status file.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>