status: Check for missing values
The YAJL_GET_NUMBER macro assumes that it is called with non-NULL
value, causing segmentation fault if this is not the case.

Assuming the status values for 'rootfs', 'bundle', 'created'
and 'cgroup-path' are not optional, this patch also shows an
error message when they are missing.

Fix: The 'detached' and 'systemd-cgroup' values were stored
as a string, not a boolean.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>