cgroup, v2: treat negative numbers as "max"
Cgroup v1 kernel doc [1] says:

> We can write "-1" to reset the *.limit_in_bytes(unlimited).

while cgroup v2 kernel documentation [2] says:

> If a controller implements an absolute resource guarantee and/or
> limit, the interface files should be named "min" and "max"
> respectively. If a controller implements best effort resource
> guarantee and/or limit, the interface files should be named "low"
> and "high" respectively.
>
> In the above four control files, the special token "max" should be
> used to represent upward infinity for both reading and writing.

Allow -1 value to still be used for v2, converting it to "max"
where it makes sense to do so.

This fixes the following error on cgroupv2:

 # crun update --pids -1 $CTID
 writing file `pids.max`: Numerical result out of range

Incidentally, `crun update --memory -1` worked fine since
"%lu" converted it to unsigned UINT64_MAX, which is
happily accepted for memory.max by the kernel.

This behavior is not documented though, does not work for all
files (e.g. pids.max) and can be changed in the future kernels.

PS all other negative numbers are still treated as is, which
might be a violation of OCI spec.

[1] linux/Documentation/admin-guide/cgroup-v1/memory.rst
[2] linux/Documentation/admin-guide/cgroup-v2.rst

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>