criu: tell CRIU to not touch the network namespace
CRIU can checkpoint and restore all namespaces. For the network
namespace either special steps are needed to re-connect the network
devices like veth back into the network namespace or, like when being
driven by Podman, the network namespace if configured externally.

For checkpoint/restore with Podman we are telling CRIU to not checkpoint
the network namespace settings and on restore we tell CRIU to use the
network namespace provided by Podman/CNI by looking at config.json if
the network namespace is defined by a 'path'.

With this the restored container will be running in the given network
namespace.

Signed-off-by: Adrian Reber <areber@redhat.com>