Save stdin, stdout, stderr pointers
To restore a container we need to reconnect stdin, stdout and stderr of
the restored process to crun. For this we read out, just after forking,
where stdin, stdout and stderr are pointing in the forked process and
store that information in crun's status file as 'external_descriptors'.
This information can in the future be used by CRIU to reconnect the
pipes of the forked process to stdin, stdout and stderr just as they
used to be.

Signed-off-by: Adrian Reber <areber@redhat.com>