criu: reconnect stdin, stdout, stderr on restore
Up until now a restored process was no longer able to output its stdout
and stderr. This uses the information stored in 'external_descriptors'
to tell CRIU to reconnect pipes back to 0, 1 and 2 if available.

With this a container restored via Podman seems to behave in the same
way as the container original started by Podman.

Container migration does not yet work, because crun creates a few
directories implicitly on container creation (like /run) and these
directories also need to be created on restore if the corresponding
file-system does not exist anymore.

So 'podman container checkpoint -l' and 'podman container restore -l'
work, but using 'checkpoint --export=...' does not work yet.

Signed-off-by: Adrian Reber <areber@redhat.com>