Recreate mountpoints during restore
If restoring a container from an unmodified root file-system certain
mountpoints have to be re-created during restore. This happens when
using Podman to restore from an exported checkpoint.

In the case of 'podman container checkpoint -l', the checkpointed
container is not destroyed and the root file-system is kept.

If restoring from an exported checkpoint using
'podman container restore --import=export.tar.gz' the root file-system
is re-created and certain mountpoints will be missing. A good example is
'/run/.containerenv'. Usually no container has the mountpoint for
'/run/.containerenv' so crun creates the mountpoint and bind mounts
something on top of it as described in config.json when the container is
created.

During restore this mountpoint has to be created just as during initial
container creation and this is what this commit now does. If the
mountpoint is not created CRIU fails during restore when trying to
remount, in this case, the corresponding bind mount.

With this it is now possible to migrate containers using Podman.

Signed-off-by: Adrian Reber <areber@redhat.com>