container: allow hooks output to file
support two custom annotations "run.oci.hooks.stdout" and
"run.oci.hooks.stderr" that allow to specify the output for hook
processes.

When the annotations are specified, the value is used as destination
file for the hooks stdout and stderr.

The file is open in append mode and it is created if it doesn't
exist.

Closes: https://github.com/containers/crun/issues/353

Signed-off-by: Giuseppe Scrivano <giuseppe@scrivano.org>