linux: handle join of a different pid namespace
handle the join an an existing pid namespace when a user namespace is
created.

To do so, we need first to attempt joining the pid namespace, once
that is done, the proc file system is mounted using fsmount and then
the user namespace is created.  The reference to the pid namespace is
injected into the container.

Requires Linux 5.3+.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>