linux: fix poststart hook
when the container is created using create+start, make sure the
poststart hook is executed inside the "start" command instead of
"create".

The OCI runtime specs say:

    The poststart hooks MUST be called after the user-specified
    process is executed but before the start operation returns. For
    example, this hook can notify the user that the container process
    is spawned.

Determining when the user-specified process is executed is not trivial
from a different process, so it can happen before the execve is
performed but it is already a very late stage in the init container
setup.

Closes: https://github.com/containers/crun/issues/406

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>