criu: tell CRIU to use the cgroup freezer
CRIU can either freeze/pause all processes in the container by going
through a loop and using ptrace (default) or use the cgroup freezer to
pause all processes in a container. This changes crun to tell CRIU the
information about the used cgroup freezer.

This needs the not yet released CRIU 3.15 which should be released in
the next weeks.

Signed-off-by: Adrian Reber <areber@redhat.com>