criu: enable checkpoint/restore in crun
Now that all functionality for the Podman checkpoint/restore test suite
has been implemented in crun, the 'checkpoint' and 'restore' commands
are no longer hidden (with the '_' prefix), but enabled to be used.

Checkpoint/restore support will not show up in crun unless CRIU 3.15 is
available (which has not yet been released).

Signed-off-by: Adrian Reber <areber@redhat.com>