libcrun: fallback to openat if openat2 returns EPERM
systemd-nspawn explictly blocks openat2 by default, relying on
implementations to fall back to older, openat code. However, it does
that by setting up a seccomp rule that returns EPERM, which libcrun is
treating as an error and thus not falling back.

Add EPERM to the list of ENOSYS/EINVAL for which openat2 errors are
ignored, and the openat fallback is attempted.

Closes: https://github.com/containers/crun/issues/545

Signed-off-by: Faidon Liambotis <paravoid@debian.org>