linux: allow arbitrary IDs with single ID userns
Create an intermediate user namespace if there is a single id mapped
inside of the container user namespace and the container is configured
to not run as root in the user namespace.

With such configuration, crun automatically creates a new user
namespace where the desired UID/GID is mapped to root in the
intermediate user namespace used to configure the container.

This enables running with an UID/GID different than root when a
rootless user has no access to multiple IDs:

$ podman run --rm --user 1000:1000 --uidmap 0:0:1 fedora:33 id
uid=1000(1000) gid=1000(1000) groups=1000(1000)

One limitation in the current implementation is that some namespaces,
like pidns are not owned by the final user namespace.

Idea copied from bubblewrap.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>