linux, init: allow mappings without root
set the UID/GID used the initial configuration to the final UID/GID
used in the container when root is not available in the container user
namespace.
Such configuration can happen for example when a single mapping is
used for rootless containers and it is mapped to the final UID/GID
used in the container.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>