Fix permission error when using both user namespaces & NOTIFY_SOCKET
If $NOTIFY_SOCKET is set, then crun will create a proxy for this socket
in the crun state directory, and bind mount it into the container.
However, the bind mounting happens after switching to the user
namespace; if UID remapping means that the container does not in fact
have permission to read that socket, then the bind mount fails.

Because the container needs permission to read _every path component_
for the source of the bind mount, and the crun state directory /run/crun
is usually root-owned and mode 0700, this actually happens _always_ when
using containers without uid 0 mapped in.

This patch fixexs the issue by first running open_tree on the
/run/crun/<container>/notify directory in the original user namespace to
_create_ the bind mount, and only putting it in the container filesystem
with move_mount after the namespace switch.

Note that when running as non-root, open_tree in the source namespace
fails - in that case, we still fall back to the bind-mount by name path
that is currently followed. In fact, this will always work, however,
since a non-root user cannot create a user namespace where UIDs are
mapped to anything _but_ the original user; and the original user can
read crun's state directory.