utils: fix symlink lookup
commit 99538d805bb62fbda6e82627d8daa0b90c97a8d0 introduced this
regression.

This issue happens only on old kernel that lack openat2.

The error happens with a command like:

$ podman run --rm -v /etc/localtime:/etc/localtime ubi8 echo "This works"
Error: open `..`: No such file or directory: OCI not found

We already have tests in place to detect this case, but they were
running on newer kernels so the code path was completely skipped.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>