ebpf: handle access(dev_name, F_OK) call correctly
F_OK does not trigger read or write access. The access type passed to
the ebpf program in that case will be zero and the initial condition

    if ((request.access & device.access) == 0)

will deny the access even if "rw" is allowed for the device file.
Instead comparing the 'masked' access type with the requested one solves
the issue:

    if ((request.access & device.access) != request.access)

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>