tests: add memhog command to init
This adds the memhog command to tests/init. This command requires a
parameter telling init how many MB of memory should be allocated.

In addition to allocation the memory and writing each page once, the
memhog command changes a memory page each 0.1 seconds. This is useful to
test crun's pre-dump support.

Signed-off-by: Adrian Reber <areber@redhat.com>