container, hooks: Proposal and support for running createContainer hooks before setting up mounts
There are use cases of `createContainer` hooks which can perform
operation on mounts before `libcrun_set_mounts` for a container

Example creating a `nfs mount` or a fresh source for `bind` mount which
was supplied to container.

Typical use-case example from `runtime-spec` gives example of running
`mount` fix-ups as `hook` when we do `createContainer`

Ref: https://github.com/opencontainers/runtime-spec/blob/main/config.md#example-2

However it will not be possible if we do `libcrun_set_mounts` before
running `createContainer` hooks following PR proposes and fixes that.

Signed-off-by: Aditya Rajan <arajan@redhat.com>