ci/gha: skip installing deps if Dockefile is used
Many tests are running inside containers, and thus they do not need any
additional packages on the host. Skip installing those if Dockerfile
is present in the test directory (for the lack of better heuristic).

The only exception from the above assumption is clang-format check,
which runs ./configure on the host, and the rest inside the container.
Fix this by adding run-tests.sh (which seems to be the standard for the
most other tests) and move ./configure to it.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>