container: attempt find_executable after setresuid
if the find_executable() call fails when running with EUID=0, ignore
any error beside ENOENT and attempt again the lookup once the process
switched to the final UID/GID used in the container.

This is visible on network file systems like NFS, where
CAP_DAC_OVERRIDE is not honored and the EUID=0 cannot read a file if
it is owned by the UID/GID in the container.

It is the same mechanism we already have in place for chdir() that
might fail before the setresuid() call.

Closes: https://github.com/containers/crun/issues/852

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>