criu: add support for different manage cgroups modes
This is motivated by checkpoint errors of containers restored with a new
ID in Podman.

Restoring a container with a new name and ID will result in a new cgroup
assigned to the container. Unfortunately CRIU restores information about
the original cgroup name. Now the container is running in one cgroup but
partially has still old cgroup information tied to it.

This adds '--manage-cgroups-mode=ignore' to crun which allows crun to
tell CRIU to completely ignore anything concerning cgroups.

With this change it is possible to checkpoint and restore a container
multiple times with changing names and IDs.

Signed-off-by: Adrian Reber <areber@redhat.com>