cgroup: removed mknod permission from default devices
This permission was added to stay consistent with runc, but even
inside runc they are not sure why to allow it. For non-root user
containers it will not change anything as one needs root to use
mknod, but for containers running as root it will no longer be
possible to create any character or block device.