linux: devices mounts should have NOEXEC and NOSUID
As in linux we cannot execute device we can safely add MS_NOEXEC
flag to the device mount. Also as it will not be executed the
flag MS_NOSUID can be added too.
Reasoning: on some platforms those flags were added either way
but on some they are missing. If it doesn't make sense to exec
then we should disallow it inside the mount just as a precausion.

Signed-off-by: Marcin Koczwara <marcin.koczwara@consult.red>