linux: fallback to tmpfs mount if umount fails
The kernel doesn't allow though to re-mount a new sysfs or proc if
there is not another mount already visible in the current mount
namespace.

When using no-pivot, we umount /sys and /proc to prevent the container
to re-mount them.  If the umount fails, just mount a tmpfs on top of
them since it will have the same effect.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>