wasmtime: always grant filesystem capability for wrkdir inside container
By default WASI programs do not have access to anything on the filesystem,
So for use cases where WASI program needs to access files on the current
path we must preopen those paths, hence by default always preopen `.` in
a `scratch` container.

For Example in below `Containerfile`, WASI program takes `inputfile` as
argument and writes output to `outputfile` where both files are present
in a path which starts with `.`.

```Dockerfile
FROM scratch
COPY inputfile /
COPY main.wasm /
CMD ["/main.wasm", "inputfile", "outputfile"]
```

Under the hood `wasmtime` preopens a directory, and make it available to
the program as a capability which can be used to open files inside that
directory. ( Using a technique based on https://github.com/musec/libpreopen )

Signed-off-by: Aditya R <arajan@redhat.com>