krun: add support for krun-sev
In addition of the regular library, libkrun provides a variant,
libkrun-sev, that turns the microVM into a TEE (Trusted Execution
Environment) using AMD SEV/SEV-ES/SEV-SNP.

This commit updates the krun handler to support using libkrun-sev in
addition of the regular libkrun. If the file "krun-sev.json" [1] is
found in the root of the OCI image, the handler will attempt to use
libkrun-sev and fail if it wasn't dlopen'ed. If the file isn't found,
the handler will attempt to use libkrun and work the same way as
before.

The handler supports that only one of the variants are present in the
system, and will only fail if the one that's missing is required to
run a particular container.

[1] oci2cw (https://github.com/virtee/oci2cw) allows to transform
regular OCI images into images that are suitable to be used with
crun+libkrun-sev.

Signed-off-by: Sergio Lopez <slp@redhat.com>