autobuild copr rpms with wasmtime support
Limited to Fedora rpms for now as wasmtime rpms don't yet build on
RHEL 8/9.

wasmtime-c-api-devel can be installed from a separate copr:
`lsm5/wasmtime` at the time of this commit, but subject to change.
The wasmtime Copr is added as an `External Repository` to
`rhcontainerbot/podman-next` so there's no need for extra configuration
in the crun repo itself.

Also reduce the dependencies installed and configure opts used in
`.copr/Makefile` as we only care about making Copr's `make_srpm` method
happy and it doesn't affect the final rpms.

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>