wasm: use wasmedge library soname in dlopen
Since wasmedge implemented library versioning in "e5af947d", use
the library's soname ("libwasmedge.so.0") instead of "libwasmedge_c.so"
to ensure crun can pin to a compatible ABI version.

Notice: the soname is shipped in wasmedge 0.11.0, before this release
the libwasmedge_c.so is unversioned.

For details, please refer to
[WasmEdge PR#1783](https://github.com/WasmEdge/WasmEdge/pull/1783)

Signed-off-by: hydai <hydai@secondstate.io>