podman,CI: skip top on priviledged container
It seems we get different result of Effective capabilities when
similar test is triggered in environment other than CI, so it seems
something is changed in underlying infrastructure of CI or podman which
crun is not aware of so we get a different output of `podman top <cid> capeff`
in CI as compared to what we get in other environment and what is
expected in tests.

Skipping this as it does not looks like any change in crun affected this
diff in output and also a test in same domain `privileged CapEff` is
already being skipped so i think skipping this should be safe.

Signed-off-by: Aditya R <arajan@redhat.com>