seccomp: use helper process to send listener fd
use a helper process to send the listener fd to the child process, in
this way the sendmssg syscall happens from an environment that is not
seccomp filtered.

The helper process shares the fd table with the main process, so there
is no need to send the fd around, and no syscalls are invoked between
the seccomp filter setup and the sendmssg, so that all syscalls can
use SCMP_ACT_NOTIFY without any risk of deadlocking the runtime itself.

Closes: https://github.com/containers/crun/issues/1002

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>