container: cache the bpf generated by seccomp
store the generated seccomp bpf by the checksum of input data in the
OCI configuration file.

The cache is maintained under $RUNDIR/.cache/seccomp.  Each file is
named after its sha256 checksum, so that we can avoid the overhead of
generating the bpf from the json configuration by calculating its
checksum and seeing if it is already cached.

If a file has the sticky bit set, then it is not evicted from the cache.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>