wasm: inherit environment variables in the WasmEdge handler
In the current implementation, the WasmEdge handler doesn't
inherit the environment variables from the current process.
Hence, the environment variables that are provided by the
containers are all dropped. This will make some information
that is passed by k8s also dropped.

This commit adds `args` and `envs` test to verify the behavior
of `std::env::args()` and `std::env::vars()`.

Signed-off-by: hydai <hydai@secondstate.io>