linux: precreate devices on the host
when a user namespace is created, the user loses the ability of
creating device nodes thus only bind mounting them from the host is
permitted.  Precreate the device nodes on the host so that we can
honor the permissions and the file mode bits specified in the OCI
spec, create the mount from the host and then bind mount them into the
user namespace.

Closes: https://github.com/containers/crun/pull/1099

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>