Don't clone self from read-only mount
Motivated by https://lwn.net/Articles/918106/
which is aiming to extend the kernel memfd API to more cleanly
separate executable from writable memfds.

But this whole "copy self to memfd" is a big mess that is only necessary
if our *underlying binary is mutable*.  Many "image based" systems
nowadays have read-only root filesystems and were never vulnerable
to the original problem.

It turns out that crun has had logic to detect a read-only bind
mount for quite a while, but only allowed it in the case where
its own special environment variable was set (signaling that
the mount was almost certainly created by crun).

But if we just tweak and reorder this logic, then it extends
to other read-only mount cases.  I believe this is safe.

Notes:

- We could also extend this to check for the immutable bit, but
  I am not aware of widespread use of that.
- There is fs-verity, but it only protects the *contents* of
  the file, and we don't want to allow a malicious container
  to mutate other properties of the file like xattrs, permission
  bits etc.
- In theory, checking for a read-only mount is racy as it
  could be remounted read-write.  There are really two cases here;
  truly read-only block devices (squashfs) and ostree-type systems.
  For the first, it can never be remounted writable.  For ostree,
  we do support `ostree admin unlock` but we always make that an
  *overmounted* `overlayfs`; we don't remount the original read-only
  bind mount.  So there's no races in either case.

Signed-off-by: Colin Walters <walters@verbum.org>