linux: mount cgroup ro on /sys bind mount fallback
when it is not possible to mount a fresh sysfs, we fall back to bind
mounting /sys.

If a /sys/fs/cgroup is not present, just mask the path.

Otherwise restrict the usage of the fallback to the case where /sys is
mounted read-only so we can take advantage of the new mount API and
use 'rro'.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>