utils: change initial size for buffer
most files that we read and that cannot be stat'ed have a very small
size thus there is no need to preallocate such a big buffer.  Also,
change its size to be one less than a power of two, so when we add the
end of line character we get a nicer round size of 1024.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>